requirements.txt
groq
langgraph
langchain-core
sentence-transformers
faiss-cpu
PyMuPDF
numpy
